feat(loop): recording + bar-snap deadzone 25% (Phase 2.E)
- startRecording : EMPTY -> RECORDING, arms capture (no buffer write yet)
- capturePadEvent : append events to main buffer (REC) with µs timestamps,
  latches recordBpm + recordStartUs on first noteOn press (invariant §23.5).
  Live monitor M8 fix Q1 : noteOn/noteOff fired immediately in ALL states
  (spec §18 "percussion fixe"), so musician hears live drumming whatever
  the engine state. Tracks _padHeldLive for stopRecording flush + B-N1/B-N2/R-N1
  audit-fix consumers. Live-sort M2 fix Q4 : insertEventSorted O(n) keeps
  _events[] sorted at all times, mergeOverdub becomes O(n+m) merge.
- stopRecording : bar-snap 25% deadzone (snap down if remainder <= 0.25*barDur,
  else round up), clamp 1..64 bars, rescale event timestamps proportionally,
  flush held pads as noteOff (via insertEventSorted, M2), M6 clamp validation
  timestamp < loopDuration, flushPendingNoteOffs (B2 self-stop), startPlayback
  (stub Phase 2.F).
- insertEventSorted : M2 helper, binary-search O(log n) + O(n) shift insertion,
  stable order for same-timestamp events.
- _overdubCount bumped uint8_t -> uint16_t pour coherence signature insertEventSorted
- ViewerSerial : new emitLoopBufferFull(channel, which) helper (m9 audit
  prerequisite) -- PRIO_LOW droppable telemetry symetrique emitArpQueueFull.
- HW gate G4 OK : record motif, live monitor audible, bar-snap math correct
  (eg 3119ms@150BPM -> 2 bars=3200ms), eventCount coherent, no stuck notes.

Spec §7, §18 + invariants §23.1 (no orphan notes) + §23.5 (immutable recordBpm)
Audit fixes : M2 live-sort + M3 velocity strict + M6 timestamp clamp + M8
live drumming + B-N1/B-N2/R-N1 _padHeldLive tracker + B2 flushPendingNoteOffs
self-stop + m9 emitLoopBufferFull

+++ src/loop/LoopEngine.h
@@ -203,7 +203,7 @@ private:
 
   // --- Overdub temp buffer (committed on tapRec during OVERDUBBING) ---
   LoopEvent        _overdubEvents[MAX_LOOP_OVERDUB_EVENTS];
-  uint8_t          _overdubCount;
+  uint16_t         _overdubCount;  // uint16_t pour signature insertEventSorted (cap=128, semantic OK)
 
   // --- Playback timeline (intégration incrémentale BPM, M-fix B1) ---
   // _playStartUs : conservé pour compat / debug (timestamp d'entrée en PLAYING), pas utilisé pour position.
